backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
python-multipart
bcrypt
python-dotenv
pydantic
httpx

# Celery & Redis
Celery
redis

# For Lambda worker
langchain
langchain-openai
langgraph
playwright
langchain-community

gnews
feedparser>=6.0.10
requests

# AWS
boto3
botocore
